Implement per-block pauses via configuration item
The `pollrate` token of the PLC configuration block header is parsed, but isn't used in the code.  To be a true poll rate, each PLC query would have to be run in parallel (i.e. a process per query), which is a significant refactoring.  At the moment we have a fixed pause at the end of each query, if in `block` refresh mode, or a fixed pause at the end of all queries if in `all` refresh mode.

Pauses are required during a scan cycle to allow clients to read/write data, by acquiring the shared memory semaphore.

We should use the `pollrate` configuration item to control end of query pauses.

If we were to set the `block` pause to zero, and set the `pollrate` in the configuration to the value that the `block` pause used to have ($`10000 \mu s`$), then when running in `block` refresh mode, it would behave identically.  The advantage though, is that we could change the pause for each block independently.  This would give us a *pseudo* pollrate.  Pseudo because the query for each block is run sequentially, so a longer pause after one block would have an effect on the *arrival time* of subsequent blocks.  Nonetheless, this is a useful configuration option.

If we only used the `pollrate` in `block` refresh mode, then the `all` refresh mode wouldn't change, and so would also behave identically.

Alternatively, we could pause after each query as well as at the end of all queries.  This would change the behaviour, but may well be acceptable.  It would mean that a pseudo pollrate could be established for a given PLC (as in `block` refresh mode), and in addition would give an extra opportunity at the end of all queries for clients to acquire the shared memory semaphore.

**N.B.:** It is still worth keeping the `block` and `all` pauses, even if they default to zero, because they can be dynamically reset at runtime via the Server Programming Interface (SPI) without stopping `pdsd`.